+++ AEWatchRenderManager/MainWindow.xaml.cs
@@ -18,8 +18,8 @@ using System.Windows.Controls;
 
 namespace AEWatchRenderManager
 {
-    // Date: Thu Apr 16 11:51:21 JST 2026
-    // Version: 1.16.16
+    // Date: Sat Apr 18 08:45:46 JST 2026
+    // Version: 1.16.17
     public partial class MainWindow : Window
     {
         public MainWindow()